+++ ltp/daemon/ltpclock.c
@@ -10,19 +10,19 @@
 									*/
 #include "ltpP.h"
 
-static int	_running(int *newValue)
+static int	_running(long *newValue)
 {
 	void	*value;
-	int	state;
+	long	state;
 
 	if (newValue)			/*	Changing state.		*/
 	{
 		value = (void *) (*newValue);
-		state = (int) sm_TaskVar(&value);
+		state = (long) sm_TaskVar(&value);
 	}
 	else				/*	Just check.		*/
 	{
-		state = (int) sm_TaskVar(NULL);
+		state = (long) sm_TaskVar(NULL);
 	}
 
 	return state;
@@ -30,7 +30,7 @@ static int	_running(int *newValue)
 
 static void	shutDown()	/*	Commands ltpclock termination.	*/
 {
-	int	stop = 0;
+	long	stop = 0;
 
 	oK(_running(&stop));	/*	Terminates ltpclock.		*/
 }
@@ -225,7 +225,7 @@ int	main(int argc, char *argv[])
 #endif
 	Sdr	sdr;
 	LtpDB	*ltpConstants;
-	int	state = 1;
+	long	state = 1;
 	time_t	currentTime;
 
 	if (ltpInit(0) < 0)
